+++ src/assets/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 4.0.0-beta.1 (2018-08-10)
+
+#### New
++ [detectEventChanges] Use it to avoid unwanted change detections. Note that, if you set it false, `chartClick` and some other event-emitters become silent.
+
+#### Change
++ [theme] now detect changes at runtime. Note that, when you change theme at runtime, the chart would be destroyed first and then initialized again.
+
 ## 4.0.0-beta.0 (2018-08-03)
 
 #### BREAKING CHANGES